Release: Spoke v11.1

+++ src/server/api/mutations/releaseCampaignNumbers.js
@@ -2,7 +2,7 @@ import { r } from "../../models";
 import cacheableData from "../../models/cacheable_queries";
 import { accessRequired } from "../errors";
 import { getConfig } from "../lib/config";
-import twilio from "../lib/twilio";
+import twilio from "../../../extensions/service-vendors/twilio";
 import ownedPhoneNumber from "../lib/owned-phone-number";
 
 export const releaseCampaignNumbers = async (_, { campaignId }, { user }) => {
@@ -33,10 +33,24 @@ export const releaseCampaignNumbers = async (_, { campaignId }, { user }) => {
   }
 
   if (service === "twilio") {
-    await twilio.deleteMessagingService(
+    const shouldRetainServices = getConfig(
+      "CAMPAIGN_PHONES_RETAIN_MESSAGING_SERVICES",
       organization,
-      campaign.messageservice_sid
+      { truthy: 1 }
     );
+
+    if (shouldRetainServices) {
+      // retain messaging services for analytics, just clear phones
+      await twilio.clearMessagingServicePhones(
+        organization,
+        campaign.messageservice_sid
+      );
+    } else {
+      await twilio.deleteMessagingService(
+        organization,
+        campaign.messageservice_sid
+      );
+    }
   }
 
   await r
